Move runs() method to live next to hollow batch
Just a little easier to find this way and more obviously not
compaction-specific.
